Moved isWinToolsImage function to a different file
