+++ frontend/packages/kubevirt-plugin/src/selectors/vm/volume.ts
@@ -1,7 +1,5 @@
 import * as _ from 'lodash';
 
-import { winToolsContainerNames } from '../../constants/vm/wintools';
-
 export const getVolumePersistentVolumeClaimName = (volume) =>
   _.get(volume, 'persistentVolumeClaim.claimName');
 export const getVolumeDataVolumeName = (volume) => _.get(volume, 'dataVolume.name');
@@ -10,8 +8,3 @@ export const getVolumeCloudInitNoCloud = (volume) => volume && volume.cloudInitN
 
 export const getVolumeContainerImage = (volume) =>
   volume && volume.containerDisk && volume.containerDisk.image;
-
-export const isWinToolsImage = (image) => {
-  const containerNames = winToolsContainerNames();
-  return Object.values(containerNames).find((winTool) => image && image.startsWith(winTool));
-};
