+++ frontend/packages/kubevirt-plugin/src/selectors/vm/winimage.ts
@@ -1,0 +1,6 @@
+import { winToolsContainerNames } from '../../constants/vm/wintools';
+
+export const isWinToolsImage = (image) => {
+  const containerNames = winToolsContainerNames();
+  return Object.values(containerNames).find((winTool) => image && image.startsWith(winTool));
+};
